Improve the support for :only option in fetch queries.
By default mongo will return the :_id field when it's not explicitly excluded.
We change this behaviour to make API simpler.

If a field in part of the :only option, it will be returned from the return data.
If a field is part of the :exclude option, it will be removed from the return
data.

:only takes precedence over :exclude